// back-end/src/entities/Autorizacao.d.ts
import type { Colaborador } from "./Colaborador.js";
import type { Area } from "./Area.js";
export declare class Autorizacao {
    id: string;
    colaborador: Colaborador;
    colaborador_id: string;
    area: Area;
    area_id: string;
    cargo_permitido: string;
    validade: Date;
}
//# sourceMappingURL=Autorizacao.d.ts.map